refactor: fix visibility and imports
