+++ Cargo.toml
@@ -4,39 +4,39 @@
 resolver = "2"
 
 members = [
-    "crates/blockifier",
-    "crates/committer",
-    "crates/committer_cli",
-    "crates/gateway",
-    "crates/mempool",
-    "crates/mempool_infra",
-    "crates/mempool_node",
-    "crates/mempool_test_utils",
-    "crates/mempool_types",
-    "crates/native_blockifier",
-    "crates/papyrus_base_layer",
-    "crates/papyrus_common",
-    "crates/papyrus_config",
-    "crates/papyrus_execution",
-    "crates/papyrus_load_test",
-    "crates/papyrus_monitoring_gateway",
-    "crates/papyrus_network",
-    "crates/papyrus_node",
-    "crates/papyrus_p2p_sync",
-    "crates/papyrus_proc_macros",
-    "crates/papyrus_protobuf",
-    "crates/papyrus_rpc",
-    "crates/papyrus_storage",
-    "crates/papyrus_sync",
-    "crates/papyrus_test_utils",
-    "crates/sequencing/papyrus_block_builder",
-    "crates/sequencing/papyrus_consensus",
-    "crates/starknet-committer",
-    "crates/starknet_api",
-    "crates/starknet_client",
-    "crates/starknet_sierra_compile",
-    "crates/task_executor",
-    "crates/tests-integration",
+  "crates/blockifier",
+  "crates/committer",
+  "crates/committer_cli",
+  "crates/gateway",
+  "crates/mempool",
+  "crates/mempool_infra",
+  "crates/mempool_node",
+  "crates/mempool_test_utils",
+  "crates/mempool_types",
+  "crates/native_blockifier",
+  "crates/papyrus_base_layer",
+  "crates/papyrus_common",
+  "crates/papyrus_config",
+  "crates/papyrus_execution",
+  "crates/papyrus_load_test",
+  "crates/papyrus_monitoring_gateway",
+  "crates/papyrus_network",
+  "crates/papyrus_node",
+  "crates/papyrus_p2p_sync",
+  "crates/papyrus_proc_macros",
+  "crates/papyrus_protobuf",
+  "crates/papyrus_rpc",
+  "crates/papyrus_storage",
+  "crates/papyrus_sync",
+  "crates/papyrus_test_utils",
+  "crates/sequencing/papyrus_block_builder",
+  "crates/sequencing/papyrus_consensus",
+  "crates/starknet_api",
+  "crates/starknet_client",
+  "crates/starknet_committer",
+  "crates/starknet_sierra_compile",
+  "crates/task_executor",
+  "crates/tests-integration",
 ]
 
 [workspace.package]
