+++ crates/committer/src/patricia_merkle_tree/external_test_utils.rs
@@ -6,16 +6,19 @@ use rand::Rng;
 use serde_json::json;
 
 use super::filled_tree::tree::{FilledTree, FilledTreeImpl};
+use super::node_data::inner_node::{EdgePathLength, PathToBottom};
 use super::node_data::leaf::{Leaf, LeafModifications, SkeletonLeaf};
 use super::original_skeleton_tree::config::OriginalSkeletonTreeConfig;
+use super::original_skeleton_tree::node::OriginalSkeletonNode;
 use super::original_skeleton_tree::tree::{OriginalSkeletonTree, OriginalSkeletonTreeImpl};
-use super::types::{NodeIndex, SortedLeafIndices};
+use super::types::{NodeIndex, SortedLeafIndices, SubTreeHeight};
 use super::updated_skeleton_tree::hash_function::TreeHashFunction;
 use super::updated_skeleton_tree::tree::{UpdatedSkeletonTree, UpdatedSkeletonTreeImpl};
 use crate::felt::Felt;
 use crate::hash::hash_trait::HashOutput;
 use crate::patricia_merkle_tree::errors::TypesError;
 use crate::storage::map_storage::MapStorage;
+use crate::storage::storage_trait::{create_db_key, StarknetPrefix, StorageKey, StorageValue};
 
 impl TryFrom<&U256> for Felt {
     type Error = TypesError<U256>;
@@ -126,4 +129,113 @@ pub async fn single_tree_flow_test<L: Leaf + 'static, TH: TreeHashFunction<L> + 
     let json_storage = &json!(filled_tree.serialize());
     result_map.insert("storage_changes", json_storage);
     serde_json::to_string(&result_map).expect("serialization failed")
+}
+
+pub fn create_32_bytes_entry(simple_val: u128) -> [u8; 32] {
+    U256::from(simple_val).to_be_bytes()
+}
+
+fn create_patricia_key(val: u128) -> StorageKey {
+    create_db_key(StarknetPrefix::InnerNode.to_storage_prefix(), &U256::from(val).to_be_bytes())
+}
+
+fn create_binary_val(left: u128, right: u128) -> StorageValue {
+    StorageValue(
+        (create_32_bytes_entry(left).into_iter().chain(create_32_bytes_entry(right))).collect(),
+    )
+}
+
+fn create_edge_val(hash: u128, path: u128, length: u8) -> StorageValue {
+    StorageValue(
+        create_32_bytes_entry(hash)
+            .into_iter()
+            .chain(create_32_bytes_entry(path))
+            .chain([length])
+            .collect(),
+    )
+}
+
+pub fn create_binary_entry(left: u128, right: u128) -> (StorageKey, StorageValue) {
+    (create_patricia_key(left + right), create_binary_val(left, right))
+}
+
+pub fn create_edge_entry(hash: u128, path: u128, length: u8) -> (StorageKey, StorageValue) {
+    (create_patricia_key(hash + path + u128::from(length)), create_edge_val(hash, path, length))
+}
+
+pub fn create_binary_skeleton_node(idx: u128) -> (NodeIndex, OriginalSkeletonNode) {
+    (NodeIndex::from(idx), OriginalSkeletonNode::Binary)
+}
+
+pub fn create_edge_skeleton_node(
+    idx: u128,
+    path: u128,
+    length: u8,
+) -> (NodeIndex, OriginalSkeletonNode) {
+    (
+        NodeIndex::from(idx),
+        OriginalSkeletonNode::Edge(
+            PathToBottom::new(path.into(), EdgePathLength::new(length).unwrap()).unwrap(),
+        ),
+    )
+}
+
+pub fn create_unmodified_subtree_skeleton_node(
+    idx: u128,
+    hash_output: u128,
+) -> (NodeIndex, OriginalSkeletonNode) {
+    (
+        NodeIndex::from(idx),
+        OriginalSkeletonNode::UnmodifiedSubTree(HashOutput(Felt::from(hash_output))),
+    )
+}
+
+pub fn create_root_edge_entry(
+    old_root: u128,
+    subtree_height: SubTreeHeight,
+) -> (StorageKey, StorageValue) {
+    // Assumes path is 0.
+    let length = SubTreeHeight::ACTUAL_HEIGHT.0 - subtree_height.0;
+    let new_root = old_root + u128::from(length);
+    let key = create_db_key(
+        StarknetPrefix::InnerNode.to_storage_prefix(),
+        &Felt::from(new_root).to_bytes_be(),
+    );
+    let value = StorageValue(
+        Felt::from(old_root)
+            .to_bytes_be()
+            .into_iter()
+            .chain(Felt::from(0_u128).to_bytes_be())
+            .chain([length])
+            .collect(),
+    );
+    (key, value)
+}
+
+pub fn create_expected_skeleton_nodes(
+    nodes: Vec<(NodeIndex, OriginalSkeletonNode)>,
+    height: u8,
+) -> HashMap<NodeIndex, OriginalSkeletonNode> {
+    let subtree_height = SubTreeHeight::new(height);
+    nodes
+        .into_iter()
+        .map(|(node_index, node)| (NodeIndex::from_subtree_index(node_index, subtree_height), node))
+        .chain([(
+            NodeIndex::ROOT,
+            OriginalSkeletonNode::Edge(
+                PathToBottom::new(0.into(), EdgePathLength::new(251 - height).unwrap()).unwrap(),
+            ),
+        )])
+        .collect()
+}
+
+impl NodeIndex {
+    /// Assumes self represents an index in a smaller tree height. Returns a node index represents
+    /// the same index in the starknet state tree as if the smaller tree was 'planted' at the lowest
+    /// leftmost node from the root.
+    pub fn from_subtree_index(subtree_index: Self, subtree_height: SubTreeHeight) -> Self {
+        let height_diff = SubTreeHeight::ACTUAL_HEIGHT.0 - subtree_height.0;
+        let offset = (NodeIndex::ROOT << height_diff) - 1.into();
+        subtree_index + (offset << (subtree_index.bit_length() - 1))
+    }
 }
