+++ crates/committer/src/patricia_merkle_tree/updated_skeleton_tree/hash_function.rs
@@ -38,5 +38,3 @@ pub trait TreeHashFunction<L: Leaf> {
         }
     }
 }
-
-
