+++ crates/committer_cli/Cargo.toml
@@ -28,6 +28,7 @@ serde_json = "1.0.116"
 serde_repr = "0.1.19"
 starknet-types-core.workspace = true
 starknet_api.workspace = true
+starknet_committer = { path = "../starknet_committer" }
 strum.workspace = true
 strum_macros.workspace = true
 thiserror.workspace = true
