+++ crates/committer_cli/src/commands.rs
@@ -1,5 +1,5 @@
-use committer::block_committer::commit::commit_block;
-use committer::block_committer::input::{Config, ConfigImpl, Input};
+use starknet_committer::block_committer::commit::commit_block;
+use starknet_committer::block_committer::input::{Config, ConfigImpl, Input};
 use tracing::info;
 use tracing::level_filters::LevelFilter;
 use tracing_subscriber::reload::Handle;
@@ -15,7 +15,7 @@ pub async fn parse_and_commit(
 ) {
     let input = parse_input(input_string).expect("Failed to parse the given input.");
     info!(
-        "Parsed committer input successfully. Original Contracts Trie Root Hash: {:?}, 
+        "Parsed committer input successfully. Original Contracts Trie Root Hash: {:?},
     Original Classes Trie Root Hash: {:?}",
         input.contracts_trie_root_hash, input.classes_trie_root_hash,
     );
@@ -32,7 +32,7 @@ pub async fn commit(input: Input<ConfigImpl>, output_path: String) {
     let output = serialized_filled_forest.forest_to_output();
     write_to_file(&output_path, &output);
     info!(
-        "Successfully committed given block. Updated Contracts Trie Root Hash: {:?}, 
+        "Successfully committed given block. Updated Contracts Trie Root Hash: {:?},
     Updated Classes Trie Root Hash: {:?}",
         output.contract_storage_root_hash, output.compiled_class_root_hash,
     );
