+++ crates/committer_cli/src/filled_tree_output/errors.rs
@@ -1,6 +1,6 @@
 use std::fmt::Debug;
 
-use committer::patricia_merkle_tree::filled_tree::errors::{
+use starknet_committer::patricia_merkle_tree::types::{
     ClassesTrieError,
     ContractsTrieError,
     StorageTrieError,
