+++ crates/committer_cli/src/parse_input/cast.rs
@@ -1,6 +1,10 @@
 use std::collections::HashMap;
 
-use committer::block_committer::input::{
+use committer::felt::Felt;
+use committer::hash::hash_trait::HashOutput;
+use committer::storage::errors::DeserializationError;
+use committer::storage::storage_trait::{StorageKey, StorageValue};
+use starknet_committer::block_committer::input::{
     ConfigImpl,
     ContractAddress,
     Input,
@@ -8,11 +12,7 @@ use committer::block_committer::input::{
     StarknetStorageValue,
     StateDiff,
 };
-use committer::felt::Felt;
-use committer::hash::hash_trait::HashOutput;
-use committer::patricia_merkle_tree::filled_tree::node::{ClassHash, CompiledClassHash, Nonce};
-use committer::storage::errors::DeserializationError;
-use committer::storage::storage_trait::{StorageKey, StorageValue};
+use starknet_committer::patricia_merkle_tree::types::{ClassHash, CompiledClassHash, Nonce};
 
 use crate::parse_input::raw_input::RawInput;
