+++ crates/committer_cli/src/tests/regression_tests.rs
@@ -2,12 +2,12 @@ use std::collections::HashMap;
 use std::fs;
 
 use clap::Error;
-use committer::block_committer::input::{ConfigImpl, Input, StarknetStorageValue};
 use committer::patricia_merkle_tree::external_test_utils::single_tree_flow_test;
-use committer::patricia_merkle_tree::original_skeleton_tree::config::OriginalSkeletonStorageTrieConfig;
-use committer::patricia_merkle_tree::updated_skeleton_tree::hash_function::TreeHashFunctionImpl;
 use serde::{Deserialize, Deserializer};
 use serde_json::{Map, Value};
+use starknet_committer::block_committer::input::{ConfigImpl, Input, StarknetStorageValue};
+use starknet_committer::hash_function::hash::TreeHashFunctionImpl;
+use starknet_committer::patricia_merkle_tree::tree::OriginalSkeletonStorageTrieConfig;
 use tempfile::NamedTempFile;
 
 use super::utils::parse_from_python::parse_input_single_storage_tree_flow_test;
