+++ crates/committer_cli/src/tests/utils/parse_from_python.rs
@@ -1,6 +1,5 @@
 use std::collections::HashMap;
 
-use committer::block_committer::input::StarknetStorageValue;
 use committer::felt::Felt;
 use committer::hash::hash_trait::HashOutput;
 use committer::patricia_merkle_tree::node_data::leaf::LeafModifications;
@@ -9,6 +8,7 @@ use committer::storage::map_storage::MapStorage;
 use committer::storage::storage_trait::{StorageKey, StorageValue};
 use ethnum::U256;
 use serde::{Deserialize, Deserializer};
+use starknet_committer::block_committer::input::StarknetStorageValue;
 
 use crate::parse_input::cast::add_unique;
 use crate::parse_input::raw_input::RawStorageEntry;
