--- crates/starknet-committer/src/starknet_patricia_merkle_tree.rs
@@ -1,4 +1,0 @@
-pub mod node;
-pub mod starknet_leaf;
-pub mod tree;
-pub mod types;
