+++ crates/starknet_committer/src/forest.rs
@@ -1,4 +1,6 @@
 pub mod filled_forest;
 pub mod forest_errors;
 pub mod original_skeleton_forest;
+#[cfg(test)]
+pub mod skeleton_forest_test;
 pub mod updated_skeleton_forest;
