+++ crates/starknet_committer/src/forest/filled_forest.rs
@@ -10,14 +10,16 @@ use committer::storage::storage_trait::Storage;
 use tokio::task::JoinSet;
 
 use crate::block_committer::input::{ContractAddress, StarknetStorageValue};
-use crate::hash::ForestHashFunction;
-use crate::starknet_forest::forest_errors::{ForestError, ForestResult};
-use crate::starknet_forest::updated_skeleton_forest::UpdatedSkeletonForest;
-use crate::starknet_patricia_merkle_tree::node::{ClassHash, CompiledClassHash, Nonce};
-use crate::starknet_patricia_merkle_tree::starknet_leaf::leaf::ContractState;
-use crate::starknet_patricia_merkle_tree::types::{
+use crate::forest::forest_errors::{ForestError, ForestResult};
+use crate::forest::updated_skeleton_forest::UpdatedSkeletonForest;
+use crate::hash_function::hash::ForestHashFunction;
+use crate::patricia_merkle_tree::leaf::leaf_impl::ContractState;
+use crate::patricia_merkle_tree::types::{
+    ClassHash,
     ClassesTrie,
+    CompiledClassHash,
     ContractsTrie,
+    Nonce,
     StorageTrie,
     StorageTrieMap,
 };
