+++ crates/starknet_committer/src/hash_function.rs
@@ -1,0 +1,3 @@
+pub mod hash;
+#[cfg(test)]
+pub mod hash_test;
