+++ crates/starknet_committer/src/lib.rs
@@ -1,0 +1,4 @@
+pub mod block_committer;
+pub mod forest;
+pub mod hash_function;
+pub mod patricia_merkle_tree;
