+++ crates/starknet_committer/src/patricia_merkle_tree.rs
@@ -1,0 +1,5 @@
+pub mod leaf;
+pub mod tree;
+pub mod types;
+#[cfg(test)]
+pub mod types_test;
