+++ crates/starknet_committer/src/patricia_merkle_tree/leaf.rs
@@ -1,4 +1,4 @@
-pub mod leaf;
+pub mod leaf_impl;
 pub mod leaf_serde;
 #[cfg(test)]
 pub mod leaf_serde_test;
